Create ue4ss-functions.mdx
honestly not sure where to put this on the side bar. it could come before Intro to Hooking, but if so then they both need reworked. could be later down as a reference type thing, but I feel like the initial examples are good to have early. and it leads into digging 101 decently so ig it's alright where it's at. or wherever you think best