Lua modding overhaul (#9)
* renamed lua-interop to lua modding

* moved lua basics to lua modding

* changed category description a bit

* renamed lua-interop to lua-modding

* added some pngs

* created page for config files

* created page for dt

* Create digging-101.mdx

* small changes to hooking-funcs

mostly just name change and then a little bit of rewording to fit it in better with the others. maybe eventually we rework this to come after UE4SS func overview, but i didn't want to stomp all over other peoples hard work

* moved lua basics to Lua Modding

* added an intro

idk if you guys want the EA disclaimer or not. i put it on my gist because it felt relevant but it might be a bit too spicy for the actual wiki idk

* renamed Creating a Lua Mod

renamed this since it's more about actually setting up the structure than creating a mod. feel free to overrule me there if you feel differently

alternatively could be "Setting Up" or "Getting Ready" or smthn to that extent

* added a tools section

i dont recall anywhere actually explaining how to install and use UE4SS on the wiki so i figured that's probably worth adding. added in the other two just for completeness sake

* Create more-digging.mdx

* Create packing-a-lua-mod.mdx

* Create ue4ss-functions.mdx

honestly not sure where to put this on the side bar. it could come before Intro to Hooking, but if so then they both need reworked. could be later down as a reference type thing, but I feel like the initial examples are good to have early. and it leads into digging 101 decently so ig it's alright where it's at. or wherever you think best

* updated translations ig?

honestly idk if these need updated or not, I just did a find and replace for "lua interop"....

* Update .gitignore

* rephrased the lua basics part

* added note + clarification

- added mod manager note
- clarified that the generated lua types can also be searched like the header files, as mentioned by Buckminsterfullerene

* moved hooking functions order

+ slight change to the beginning

this could maybe be reworked more now that we have the initial UE4SS functions page, but i dont want to be the one to call that shot

* added read-only note

* moved position up

* deleted packing a lua mod

* moved up + reworked a bit

rewrote a bit of this since it'll technically be their first exposure to UE4SS functions now